+++ lib/timer_face/PieTimer.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:vibration/vibration.dart';
-import 'dart:io';
 import 'CustomTimerPainter.dart';
-import 'Util.dart';
+import '../util/Util.dart';
 import '../DAO.dart';
 
 enum PieTimerStatus { none, playing, paused, cancelled }
@@ -36,8 +34,8 @@ class _PieTimerState extends State<PieTimer> with TickerProviderStateMixin {
       ..addStatusListener((animationStatus) {
         // listens for changes to the animation to update the timer status
         if (animationStatus == AnimationStatus.dismissed) {
-          _vibrateAlert(5);
-          getDialog(context, "Timer Complete", "The timer is finished.");
+          vibrateAlert(5);
+          getTextDialog(context, "Timer Complete", "The timer is finished.");
           _switchStatus(PieTimerStatus.none);
         }
       });
@@ -51,7 +49,7 @@ class _PieTimerState extends State<PieTimer> with TickerProviderStateMixin {
     return Scaffold(
         backgroundColor: Colors.transparent,
         body: AnimatedBuilder(
-            animation: getUpdate(context),
+            animation: (getUpdate(context)),
             builder: (context, child) =>
                 // buildStack pushes all static values to be positioned and separately adds
                 // elements that need to be rebuilt every time the animation controller changes
@@ -200,16 +198,6 @@ class _PieTimerState extends State<PieTimer> with TickerProviderStateMixin {
   // verifies that a reset is possible (e.g., that the timer is actually running)
   bool _checkTimerStatusReset() {
     return (_status == PieTimerStatus.playing); 
-  }
-
-
-  // run the vibration for the alert
-  void _vibrateAlert(int vibrationRepetition) {
-    // run the vibration
-    for (var i = 0; i < vibrationRepetition; i++) {
-      Vibration.vibrate(duration: 150, amplitude: 250);
-      sleep(const Duration(milliseconds: 200));
-    }
   }
 
   // update after setTime
